+++ packages/components/rating/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @jamsr-ui/rating
 
+## 0.0.3
+
+### Patch Changes
+
+- fix issues
+
+## 0.0.2
+
+### Patch Changes
+
+- fix types
+
 ## 0.0.1
 
 ### Patch Changes